# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Serenibase is a workspace management platform that layers guardrails on PostgreSQL tables and permissions.\nIt centralizes workspace/base/table creation, membership, and asset/connectivity workflows along with RBAC controls.",
        "title": "Serenibase API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support Team",
            "email": "support@serenibase.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/base/{id}/bulk-add-members": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds multiple users to the provided base according to their requested base roles.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Workspace"
                ],
                "summary": "Bulk add members to a base",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional client-generated request trace ID",
                        "name": "X-Request-ID",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Base ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Bulk membership payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkAddBaseMembersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bulk additions reported",
                        "schema": {
                            "$ref": "#/definitions/dto.BulkAddMembersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request — missing fields",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized — invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — insufficient privileges",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workspace/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns every workspace visible to the tenant including metadata.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Workspace"
                ],
                "summary": "List all workspaces",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional client-generated request trace ID",
                        "name": "X-Request-ID",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workspaces retrieved successfully (data contains workspace list)",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized — invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — insufficient privileges",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workspace/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Persists a new workspace entity for the tenant, storing the title and metadata provided by the requester.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Workspace"
                ],
                "summary": "Create a workspace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional client-generated request trace ID",
                        "name": "X-Request-ID",
                        "in": "header"
                    },
                    {
                        "description": "Workspace payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateWorkspaceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Workspace created",
                        "schema": {
                            "$ref": "#/definitions/dto.WorkspaceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request — missing title or workspace data",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized — invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — insufficient privileges",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workspace/{id}/bases": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns every base the workspace owns, including access-level meta for the requesting user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Workspace"
                ],
                "summary": "List bases for a workspace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional client-generated request trace ID",
                        "name": "X-Request-ID",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bases retrieved",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BaseResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized — invalid authentication",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — missing access level",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found — workspace missing",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/workspace/{id}/bulk-add-members": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates multiple workspace access records using the memberships described in the request.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Workspace"
                ],
                "summary": "Bulk add members to workspace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional client-generated request trace ID",
                        "name": "X-Request-ID",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Target workspace ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Array of member and membership definitions",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkAddMembersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bulk member addition results",
                        "schema": {
                            "$ref": "#/definitions/dto.BulkAddMembersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request — invalid payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized — invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — not allowed to bulk add",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workspace/{id}/tables": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the table definitions owned by the workspace so front-end can render the schema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Workspace"
                ],
                "summary": "Get tables scoped to a workspace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional client-generated request trace ID",
                        "name": "X-Request-ID",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tables listed",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.TableResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized — invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — insufficient access",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found — workspace missing",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.BulkAddBaseMembersRequest": {
            "type": "object",
            "required": [
                "members"
            ],
            "properties": {
                "members": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.BulkMemberRequest"
                    }
                }
            }
        },
        "dto.BulkAddMembersRequest": {
            "type": "object",
            "required": [
                "members"
            ],
            "properties": {
                "members": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.BulkMemberRequest"
                    }
                }
            }
        },
        "dto.BulkAddMembersResponse": {
            "type": "object",
            "properties": {
                "failed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MemberAddFailure"
                    }
                },
                "success": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.BulkMemberRequest": {
            "type": "object",
            "required": [
                "memberships",
                "user_id"
            ],
            "properties": {
                "memberships": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.MembershipRequest"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateWorkspaceRequest": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/models.ErrorDetail"
                },
                "message": {
                    "type": "string",
                    "example": "Validation failed"
                },
                "meta": {
                    "$ref": "#/definitions/models.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "meta": {
                    "$ref": "#/definitions/models.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your Bearer token in the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}